extensions/publisher-command-center-otel/requirements.txt
# requirements.txt auto-generated by Posit Publisher
# using /Users/me/Projects/connect-extensions/extensions/content-manager/.venv/bin/python
cachetools==5.5.1
fastapi==0.115.6
opentelemetry-api==1.39.0
opentelemetry-sdk==1.39.0
opentelemetry-instrumentation-fastapi==0.60b0
opentelemetry-exporter-otlp-proto-http==1.39.0
posit_sdk @ git+https://github.com/posit-dev/posit-sdk-py.git@main
